H.Res. 97 Engrossed in House
Providing for consideration of the bill (H.R. 185) to terminate the requirement imposed by the Director of the Centers for Disease Control and Prevention for proof of COVID-19 vaccination for foreign travelers, and for other purposes; providing for consideration of the joint resolution (H.J. Res. 24) disapproving the action of the District of Columbia Council in approving the Local Resident Voting Rights Amendment Act of 2022; and providing for consideration of the joint resolution (H.J. Res. 26) disapproving the action of the District of Columbia Council in approving the Revised Criminal Code Act of 2022.

Version:  Engrossed in House (eh)
Date:     2023-02-07
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres97eh/xml/BILLS-118hres97eh.xml

Sponsored-By: L000600
Roll-Call: House 118-1-108 2023-02-07 Passed 217-208
Roll-Call: House 118-1-109 2023-02-07 Passed 217-208