H.Res. 166 Reported in House
Providing for consideration of the bill (H.R. 347) to require the Executive Office of the President to provide an inflation estimate with respect to Executive orders with a significant effect on the annual gross budget, and for other purposes, and providing for consideration of the joint resolution (H.J. Res. 30) providing for congressional disapproval under chapter 8 of title 5, United States Code, of the rule submitted by the Department of Labor relating to "Prudence and Loyalty in Selecting Plan Investments and Exercising Shareholder Rights".

Version:  Reported in House (rh)
Date:     2023-02-27
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres166rh/xml/BILLS-118hres166rh.xml

Sponsored-By: B001248