H.Res. 199 Reported in House
Providing for consideration of the bill (H.R. 140) to amend title 5, United States Code, to prohibit Federal employees from advocating for censorship of viewpoints in their official capacity, and for other purposes; providing for consideration of the joint resolution (H.J. Res. 27) providing for congressional disapproval under chapter 8 of title 5, United States Code, of the rule submitted by the Department of the Army, Corps of Engineers, Department of Defense and the Environmental Protection Agency relating to "Revised Definition of 'Waters of the United States'"; and providing for consideration of the bill (S. 619) to require the Director of National Intelligence to declassify information relating to the origin of COVID-19, and for other purposes.

Version:  Reported in House (rh)
Date:     2023-03-07
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres199rh/xml/BILLS-118hres199rh.xml

Sponsored-By: M001184