H.Res. 298 Reported in House
Providing for consideration of the bill (H.R. 734) to amend the Education Amendments of 1972 to provide that for purposes of determining compliance with title IX of such Act in athletics, sex shall be recognized based solely on a person's reproductive biology and genetics at birth, and providing for consideration of the joint resolution (H.J. Res. 42) disapproving the action of the District of Columbia Council in approving the Comprehensive Policing and Justice Reform Amendment Act of 2022.

Version:  Reported in House (rh)
Date:     2023-04-17
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres298rh/xml/BILLS-118hres298rh.xml

Sponsored-By: H001093