H.Res. 699 Reported in House
Providing for consideration of the bill (H.R. 1130) to repeal restrictions on the export and import of natural gas; providing for consideration of the resolution (H. Res. 684) condemning the actions of Governor of New Mexico, Michelle Lujan Grisham, for subverting the Second Amendment to the Constitution and depriving the citizens of New Mexico of their right to bear arms; and providing for consideration of the bill (H.R. 5525) making continuing appropriations for fiscal year 2024, and for other purposes.

Version:  Reported in House (rh)
Date:     2023-09-19
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres699rh/xml/BILLS-118hres699rh.xml

Sponsored-By: R000610