H.Res. 712 Reported in House
Providing for consideration of the bill (H.R. 4365) making appropriations for the Department of Defense for the fiscal year ending September 30, 2024, and for other purposes; providing for consideration of the bill (H.R. 1130) to repeal restrictions on the export and import of natural gas; and providing for consideration of the resolution (H. Res. 684) condemning the actions of Governor of New Mexico, Michelle Lujan Grisham, for subverting the Second Amendment to the Constitution and depriving the citizens of New Mexico of their right to bear arms.

Version:  Reported in House (rh)
Date:     2023-09-21
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres712rh/xml/BILLS-118hres712rh.xml

Sponsored-By: C001053
Roll-Call: House 118-1-402 2023-09-21 Passed 216-202
Roll-Call: House 118-1-403 2023-09-21 Failed 212-216