H.Res. 1071 Reported in House
Providing for consideration of the bill (H.R. 6276) to direct the Administrator of General Services and the Director of the Office of Management and Budget to identify the utilization rate of certain public buildings and federally-leased space, and for other purposes, and providing for consideration of the resolution (H. Res. 1065) denouncing the Biden administration's immigration policies.

Version:  Reported in House (rh)
Date:     2024-03-11
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres1071rh/xml/BILLS-118hres1071rh.xml

Sponsored-By: M001184