H.Res. 1085 Reported in House
Providing for consideration of the bill (H.R. 1023) to repeal section 134 of the Clean Air Act, relating to the greenhouse gas reduction fund; providing for consideration of the bill (H.R. 1121) to prohibit a moratorium on the use of hydraulic fracturing; providing for consideration of the bill (H.R. 6009) to require the Director of the Bureau of Land Management to withdraw the proposed rule relating to fluid mineral leases and leasing process, and for other purposes; providing for consideration of the concurrent resolution (H. Con. Res. 86) expressing the sense of Congress that a carbon tax would be detrimental to the United States economy; providing for consideration of the resolution (H. Res. 987) denouncing the harmful, anti-American energy policies of the Biden administration, and for other purposes; and providing for consideration of the bill (H.R 7023) to amend section 404 of the Federal Water Pollution Control Act to codify certain regulatory provisions relating to nationwide permits for dredged or fill material, and for other purposes.

Version:  Reported in House (rh)
Date:     2024-03-19
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres1085rh/xml/BILLS-118hres1085rh.xml

Sponsored-By: R000610