H.Res. 1149 Engrossed in House
Providing for consideration of the bill (H.R. 6323) to modify the availability of certain waiver authorities with respect to sanctions imposed with respect to the financial sector of Iran, and for other purposes; providing for consideration of the resolution (H. Res. 1143) condemning Iran's unprecedented drone and missile attack on Israel; providing for consideration of the bill (H.R. 4691) to provide for congressional review of actions to terminate or waive sanctions imposed with respect to Iran; providing for consideration of the bill (H.R. 5947) to provide for the rescission of certain waivers and licenses relating to Iran, and for other purposes; providing for consideration of the bill (H.R. 6046) to designate Ansarallah as a foreign terrorist organization and impose certain sanctions on Ansarallah, and for other purposes; and providing for consideration of the bill (H.R. 4639) to amend section 2702 of title 18, United States Code, to prevent law enforcement and intelligence agencies from obtaining subscriber or customer records in exchange for anything of value, to address communications and records in the possession of intermediary internet service providers, and for other purposes.

Version:  Engrossed in House (eh)
Date:     2024-04-16
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres1149eh/xml/BILLS-118hres1149eh.xml

Sponsored-By: R000610
Roll-Call: House 118-2-124 2024-04-16 Passed 212-208
Roll-Call: House 118-2-125 2024-04-16 Passed 214-208