H.Res. 1194 Engrossed in House
Providing for consideration of the bill (H.R. 6192) to amend the Energy Policy and Conservation Act to prohibit the Secretary of Energy from prescribing any new or amended energy conservation standard for a product that is not technologically feasible and economically justified, and for other purposes; providing for consideration of the bill (H.R. 7109) to require a citizenship question on the decennial census, to require reporting on certain census statistics, and to modify apportionment of Representatives to be based on United States citizens instead of all persons; providing for consideration of the joint resolution (H.J. Res. 109) providing for congressional disapproval under chapter 8 of title 5, United States Code, of the rule submitted by the Securities and Exchange Commission relating to "Staff Accounting Bulletin No. 121"; and providing for consideration of the bill (H.R. 2925) to amend the Omnibus Budget Reconciliation Act of 1993 to provide for security of tenure for use of mining claims for ancillary activities, and for other purposes.

Version:  Engrossed in House (eh)
Date:     2024-05-07
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres1194eh/xml/BILLS-118hres1194eh.xml

Sponsored-By: H001093
Roll-Call: House 118-2-180 2024-05-07 Passed 204-200
Roll-Call: House 118-2-181 2024-05-07 Passed 205-199