H.Res. 1227 Engrossed in House
Providing for consideration of the bill (H.R. 8369) to provide for the expeditious delivery of defense articles and defense services for Israel and other matters; providing for consideration of the bill (H.R. 7530) to limit youth offender status in the District of Columbia to individuals 18 years of age or younger, to direct the Attorney General of the District of Columbia to establish and operate a publicly accessible website containing updated statistics on juvenile crime in the District of Columbia, to amend the District of Columbia Home Rule Act to prohibit the Council of the District of Columbia from enacting changes to existing criminal liability sentences, and for other purposes; providing for consideration of the bill (H.R. 7343) to amend the Immigration and Nationality Act to provide for the detention of certain aliens who commit assault against law enforcement officers; providing for consideration of the bill (H.R. 8146) to require a report by the Attorney General on the impact the border crisis is having on law enforcement at the Federal, State, local, and Tribal level; providing for consideration of the bill (H.R. 7581) to require the Attorney General to develop reports relating to violent attacks against law enforcement officers, and for other purposes; providing for consideration of the bill (H.R. 354) to amend title 18, United States Code, to improve the Law Enforcement Officer Safety Act and provisions relating to the carrying of concealed weapons by law enforcement officers, and for other purposes; providing for consideration of the resolution (H. Res. 1213) a resolution regarding violence against law enforcement officers; and providing for consideration of the resolution (H. Res. 1210) condemning the Biden border crisis and the tremendous burdens law enforcement officers face as a result.

Version:  Engrossed in House (eh)
Date:     2024-05-15
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres1227eh/xml/BILLS-118hres1227eh.xml

Sponsored-By: L000600
Roll-Call: House 118-2-198 2024-05-15 Passed 212-201
Roll-Call: House 118-2-199 2024-05-15 Passed 212-200