H.Res. 1341 Reported in House
Providing for consideration of the bill (H.R. 8281) to amend the National Voter Registration Act of 1993 to require proof of United States citizenship to register an individual to vote in elections for Federal office, and for other purposes; providing for consideration of the joint resolution (H.J. Res. 165) providing for congressional disapproval under chapter 8 of title 5, United States Code, of the rule submitted by the Department of Education relating to "Nondiscrimination on the Basis of Sex in Education Programs or Activities Receiving Federal Financial Assistance"; providing for consideration of the bill (H.R. 8772) making appropriations for the Legislative Branch for the fiscal year ending September 30, 2025, and for other purposes; providing for consideration of the bill (H.R. 7700) to prohibit the Secretary of Energy from prescribing or enforcing energy conservation standards for dishwashers that are not cost-effective or technologically feasible, and for other purposes; and providing for consideration of the bill (H.R. 7637) to prohibit the Secretary of Energy from prescribing or enforcing energy conservation standards for refrigerators, refrigerator-freezers, and freezers that are not cost-effective or technologically feasible, and for other purposes.

Version:  Reported in House (rh)
Date:     2024-07-09
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres1341rh/xml/BILLS-118hres1341rh.xml

Sponsored-By: B001248
Roll-Call: House 118-2-338 2024-07-09 Passed 194-186
Roll-Call: House 118-2-339 2024-07-09 Passed 205-201