H.Res. 1455 Reported in House
Providing for consideration of the bill (H.R. 3724) to amend the Higher Education Act of 1965 to prohibit recognized accrediting agencies and associations from requiring, encouraging, or coercing institutions of higher education to meet any political litmus test or violate any right protected by the Constitution as a condition of accreditation; providing for consideration of the bill (H.R. 4790) to amend the Federal securities laws with respect to the materiality of disclosure requirements, to establish the Public Company Advisory Committee, and for other purposes; providing for consideration of the bill (H.R. 5179) to require the maintenance of the country of origin markings for imported goods produced in the West Bank or Gaza, and for other purposes; providing for consideration of the bill (H.R. 5339) to amend the Employee Retirement Income Security Act of 1974 to specify requirements concerning the consideration of pecuniary and non-pecuniary factors, and for other purposes; providing for consideration of the bill (H.R. 5717) to provide that sanctuary jurisdictions that provide benefits to aliens who are present in the United States without lawful status under the immigration laws are ineligible for Federal funds intended to benefit such aliens; providing for consideration of the bill (H.R. 7909) to amend the Immigration and Nationality Act to provide that aliens who have been convicted of or who have committed sex offenses or domestic violence are inadmissible and deportable; and providing for consideration of the joint resolution (H.J. Res. 136) providing for congressional disapproval under chapter 8 of title 5, United States Code, of the rule submitted by the Environmental Protection Agency relating to "Multi-Pollutant Emissions Standards for Model Years 2027 and Later Light-Duty and Medium-Duty Vehicles".

Version:  Reported in House (rh)
Date:     2024-09-17
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres1455rh/xml/BILLS-118hres1455rh.xml

Sponsored-By: H001093