H.Res. 1486 Reported in House
Providing for consideration of the bill (H.R. 3334) to provide for the imposition of sanctions on members of the National Communist Party Congress of the People's Republic of China, and for other purposes; providing for consideration of the bill (H.R. 8205) to amend the Omnibus Crime Control and Safe Streets Act of 1968 to provide that Byrne grant funds may be used for public safety report systems, and for other purposes; providing for consideration of the bill (H.R. 8790) to expedite under the National Environmental Policy Act of 1969 and improve forest management activities on National Forest System lands, on public lands under the jurisdiction of the Bureau of Land Management, and on Tribal lands to return resilience to overgrown, fire-prone forested lands, and for other purposes; providing for consideration of the resolution (H. Res. 1469) ensuring accountability for key officials in the Biden-Harris administration responsible for decisionmaking and execution failures throughout the withdrawal from Afghanistan; and for other purposes.

Version:  Reported in House (rh)
Date:     2024-09-23
Congress: 118

Source: https://www.govinfo.gov/content/pkg/BILLS-118hres1486rh/xml/BILLS-118hres1486rh.xml

Sponsored-By: B001248